# backend/progresstracker/openapi.json
{
  "components" : {
    "schemas" : {
      "AuthResponse" : {
        "properties" : {
          "token" : {
            "type" : "string"
          }
        },
        "required" : [
          "token"
        ],
        "type" : "object"
      },
...
      "LoginRequest" : {
        "properties" : {
          "email" : {
            "maxLength" : 255,
            "minLength" : 1,
            "type" : "string"
          },
          "password" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [
          "email",
          "password"
        ],
        "type" : "object"
      },
      "ProblemDetail" : {
        "properties" : {
          "detail" : {
            "type" : "string"
          },
          "errors" : {
            "additionalProperties" : {
              "type" : "string"
            },
            "description" : "Per-field validation messages. Present on validation failures only.",
            "type" : "object"
          },
          "instance" : {
            "format" : "uri",
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "format" : "uri",
            "type" : "string"
          }
        }
      },
...
      "RegisterRequest" : {
        "properties" : {
          "email" : {
            "format" : "email",
            "maxLength" : 255,
            "minLength" : 1,
            "type" : "string"
          },
          "password" : {
            "maxLength" : 72,
            "minLength" : 8,
            "type" : "string"
          },
          "timeZone" : {
            "maxLength" : 64,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [
          "email",
          "password"
        ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Habits, completions, streaks, XP, and achievements.",
    "title" : "Progress Tracker API",
    "version" : "1.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/auth/login" : {
      "post" : {
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            },
            "description" : "Logged in; returns a bearer token"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The request body is malformed or fails validation"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Invalid credentials"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Too many failed sign-ins for this email or from this address",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds to wait before trying again",
                "schema" : {
                  "type" : "integer"
                },
                "style" : "simple"
              }
            }
          }
        },
        "security" : [ ],
        "tags" : [
          "Authentication"
        ]
      }
    },
    "/api/auth/logout" : {
      "post" : {
        "description" : "Ends every session the caller has, on every device: every token issued to them before this call stops working. Sign in again for a new one.",
        "operationId" : "logout",
        "responses" : {
          "204" : {
            "description" : "Signed out"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Missing or invalid bearer token"
          }
        },
        "security" : [
          {
            "bearerAuth" : [ ]
          }
        ],
        "summary" : "Sign out everywhere",
        "tags" : [
          "Authentication"
        ]
      }
    },
    "/api/auth/register" : {
      "post" : {
        "operationId" : "register",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            },
            "description" : "Registered; returns a bearer token"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The request body is malformed or fails validation"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Email already used"
          }
        },
        "security" : [ ],
        "tags" : [
          "Authentication"
        ]
      }
    },
...
  },
  "security" : [
    {
      "bearerAuth" : [ ]
    }
  ]
}